Django models Refactoring for better readability
organized models used by different tests in separate
folders
