+++ frameworks/python/django/src/testApp/models/__init__.py
@@ -1,0 +1,29 @@
+from .crud import Person
+from .fields import Server,Event
+from .relationships import (
+    Book,
+    Author,
+    Article,
+    Reporter,
+    HQAddress,
+    Company
+)
+from .inheritance import (
+    CommonInfo,
+    Worker,
+    Unmanaged,
+    Student,
+    Post,
+    Attachment,
+    Comment,
+    Link,
+    Chef,
+    Place,
+    Rating,
+    Restaurant,
+    MyPerson,
+    OrderedPerson,
+    Product,
+    Review,
+    ProductReview
+)
