+++ frameworks/python/django/src/testApp/models/crud.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+class Person(models.Model):
+    first_name = models.CharField(max_length=100)
+    last_name = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.first_name
